requirements.txt
aiogram==3.13.1
playwright==1.49.0
SQLAlchemy==2.0.36
aiofiles==24.1.0
python-dotenv==1.0.1
